Reorganize GUI and UI files into appropriate packages
There are too many files in the main package.

The UI files and GUI files sometimes take times to navigate and manage.

ErrorMessage, UI, and Parser all handle user command processing and reply
generation. These classes are now grouped under the `ui` package.

Main, MainWindow, and DialogBox are responsible for creating the graphical
interface of the chatbot. These classes are now grouped under the `gui` package.

This reorganize improves the structure by grouping classes with the same purpose. making it easier to manage and navigate.